Revert "Feature/US-0002"

+++ social-meli/src/main/java/com/api/social_meli/model/Identifiable.java
@@ -1,5 +1,5 @@
 package com.api.social_meli.model;
 
 public interface Identifiable {
-    int getId();
+    Long getId();
 }
